refactor mutations
