+++ api/app/GraphQL/Mutations/SetPoolCandidateSearchRequestStatus.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\GraphQL\Mutations;
+
+use App\Enums\TalentRequestStatus;
+use App\Models\PoolCandidateSearchRequest;
+
+final class SetPoolCandidateSearchRequestStatus
+{
+    public function __invoke($_, array $args): PoolCandidateSearchRequest
+    {
+        $searchRequest = PoolCandidateSearchRequest::findOrFail($args['id']);
+
+        if ($args['status'] === TalentRequestStatus::IN_PROGRESS->name) {
+            $searchRequest->progress($args['inProgressDetails'], $args['followUpDate'] ?? null);
+        } elseif ($args['status'] === TalentRequestStatus::CLOSED->name) {
+            $searchRequest->close($args['closedDetails']);
+        }
+
+        return $searchRequest;
+    }
+}
